requirements.txt
coverage==6.2
asgiref==3.4.1
Django==4.0
djangorestframework==3.12.4
pytz==2021.3
setuptools==57.0.0
sqlparse==0.4.2
tzdata==2021.5
wheel==0.36.2